Allow /etc/ltsp/{passwd,group} overrides, e.g. for LDAP (#16)

+++ ltsp/common/initrd/35-initrd.sh
@@ -40,9 +40,9 @@ initrd_main() {
     test -f "$_DST_DIR/etc/ltsp/ssh_known_hosts" ||
         rw sed "s/^/server /" /etc/ssh/ssh_host_*_key.pub > \
             "$_DST_DIR/etc/ltsp/ssh_known_hosts"
-    # Copy server passwd and group
-    re cp -a /etc/passwd "$_DST_DIR/etc/ltsp/"
-    re cp -a /etc/group "$_DST_DIR/etc/ltsp/"
+    # Copy server passwd and group; allow overrides for LDAP etc
+    test -f /etc/ltsp/passwd || re cp -a /etc/passwd "$_DST_DIR/etc/ltsp/"
+    test -f /etc/ltsp/group || re cp -a /etc/group "$_DST_DIR/etc/ltsp/"
     # Copy epoptes keys; but provide for a future override
     if [ "$IGNORE_EPOPTES" != "1" ] && [ -f /etc/epoptes/server.crt ]; then
         re cp -a /etc/epoptes/server.crt "$_DST_DIR/etc/ltsp/epoptes.crt"
